Revert "video: bcm2708_fb: Try allocating on the ARM and passing to VPU"
This reverts commit ca36c709fce57e8023d2b8b354376bf161601a49.

The driver tries a cma_alloc to avoid using gpu_mem, but should
that fail the core code is logging an error with no easy way to
test whether it will succeed or fail first.

Revert until we either totally give up on gpu_mem and increase
CMA always, or find a way to try an allocation.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>